Move added API types helpers to dedicated package
Move the helper functions for new API types (ExtendedObjectMeta and
children) to a dedicate package to avoid changing the API anytime
a helper implementation changes.

Helper functions dealing with API types which were already there
are not moved to keep the PR focused, that is left for the future.
